+++ packages/nuxt/src/runtime/server/plugins/armor.ts
@@ -1,5 +1,6 @@
 import { createArmor } from 'ai-armor'
-import { defineNitroPlugin, useRuntimeConfig } from 'nitropack/runtime'
+import { defineNitroPlugin } from 'nitropack/runtime'
+import { useRuntimeConfig } from '#imports'
 import { initArmor } from '../utils/armor'
 
 export default defineNitroPlugin(() => {